refactor : Image 크기 수정 및 필요없는 코드 삭제
